backend/requirements.txt
flask
python-dotenv
psycopg2-binary
flask-sqlalchemy
Flask-Migrate